basic/alloc-util: introduce heap-optional newa_safe()
The bare newa() macro will fatally assert at runtime if the
requested allocation exceeds ALLOCA_MAX, making it very fragile
and arguably inappropriate to use in all but the most controlled
of circumstances.

This newa_safe() variant instead uses a plain heap allocation to
satisfy the request if ALLOCA_MAX is exceeded.  Callers are
expected to use this in a more regular new() heap-allocation
fashion along with _cleanup_newa_-qualified variables.  This
cleanup handler will free the memory when heap-allocated or do
nothing when stack-allocated.

The way this works is by enlarging the requested allocation
slightly for accomdating a free-function pointer.  When the stack
was used, the free-function pointer is set to NULL.  When the
heap was used, it's set to free().  The actual pointer assigned
to the _cleanup_newa_-qualified variable and returned by
newa_safe() points immediately after where this free-function
pointer is kept, and the caller doesn't need to concern
themselves with whether the allocation occurred on-stack or
on-heap.

This allows quick on-stack allocations when small, without
abruptly exploding should the allocation exceed ALLOCA_MAX for
whatever reason.

Note there are existing call sites where newa() is utilized
directly with caller-provided sizes in public interfaces like
sd_journal_sendv().  Such scenarios are good candidates for using
this more robust newa_safe() variant without losing the
efficiency of newa() for the expected common case.