sd-journal: use newa_safe() in sd_journal_sendv()
This is public API and we can't guarantee caller-supplied values
for n will always be safe for newa() to not explode in either its
assert on ALLOCA_MAX or simply smash the stack if compiled
with asserts disabled.

Instead use newa_safe() which will gracefully resort to the heap
if necessary, but will still use the stack when possible.